requirements.txt
analytics-python
Faker